baichuan2

+++ python/llm/src/bigdl/llm/transformers/models/baichuan.py
@@ -318,7 +318,7 @@ def baichuan_attention_forward_7b_origin(
     if not output_attentions:
         attn_weights = None
 
-    return attn_output, attn_weights, past_key_value
+    return attn_output.to(hidden_states.dtype), attn_weights, past_key_value
 
 
 def baichuan_attention_forward_13b(
@@ -536,4 +536,4 @@ def baichuan_attention_forward_13b_origin(
     if not output_attentions:
         attn_weights = None
 
-    return attn_output, attn_weights, past_key_value
+    return attn_output.to(hidden_states.dtype), attn_weights, past_key_value
